Fixed addevent route to work with mongoose
Removed date and type fields from the form in addevent.jade. Changed the URL from / to /addevent in the post request in events.js

Added redirect to post request in events.js

Added date and type to the event model, and to the form in the addevent view. POST request now saves name, date and type to the database.

Removed comment from events.js with the current branch name